General changes: restructured the entire resume. Remvoed the enumitem package. Added the tabu package. Created titleformating for both the section and paragraph headings. Changed the Hobbies section to "Interests & Activities" Changed font size of some of the headings.
Heading:
Centered the heading.

Objective section completely removed.

Education section:
Removed the Minor
Moved the location to right next to the college. Italicized the location.

Skills Section:
Updated skills section from a list to a stack.
Added relative years of use and proficiency.

Exprience Section:
Moved the company to the top and made it, it's own heading.
added more titles to positions.
Switched order of department and position.
changed the lhric website to be underlined.
bolded the title, Italicized the department, and bolded the time frame.
Changed the description for a majority of the positions.

Projects section:
Removed a project or two.
fixed some speeling mistakes.

Interests & Activities section:
Included position with the RIT Juggling club.
Changed the description of the RIT Juggling Club.
Changed the description of the RIT Computer Science House.